chore: slight refactor, moving Map to map.rs
