+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "clashmap"
-version = "1.0.0"
+version = "1.0.1"
 authors = ["Conrad Ludgate <conradludgate@gmail.com>", "Acrimon <joel.wejdenstal@gmail.com>"]
 edition = "2021"
 rust-version = "1.70"
